requirements.txt
tensorflow-gpu>=2.0.0-alpha0
numpy>=1.16.3
tfds-nightly>=1.0.2.dev201905010105
tqdm>=4.31.1
matplotlib>=3.0.3